better provider for database. Puppet resources compatible

+++ lib/puppet/provider/influxdb_database/ruby.rb
@@ -1,31 +1,53 @@
 Puppet::Type.type(:influxdb_database).provide(:ruby) do
   desc 'Manages InfluxDB databases.'
 
-commands :influx_cli => '/opt/influxdb/influx'
+  commands :influx_cli => '/opt/influxdb/influx'
 
-def get_database_list()
-  begin
-    output = influx_cli(['-execute', 'show databases'])
-  rescue Puppet::ExecutionFailure => e
-    Puppet.debug("#get_database_list had an error -> #{e.inspect}")
-    return nil
+  def initialize(value={})
+    super(value)
+    @property_flush = {}
   end
-  databases = output.split("\n")[2..-1]
-  Puppet.debug("#get_database_list output -> #{databases}")
-  databases
-end
 
-def exists?
-  dbs = get_database_list()
-  dbs.include? resource[:name]
-end
+  def self.instances
+    begin
+      output = influx_cli(['-execute', 'show databases'])
+    rescue Puppet::ExecutionFailure => e
+      return nil
+    end
+    databases = output.split("\n")[3..-1]
+    databases.collect do |name|
+      new(:name    => name,
+        :ensure  => :present,
+      )
+    end
+  end
 
-def create
-  influx_cli(['-execute', "create database #{resource[:name]}"].compact)
-end
+  def self.prefetch(resources)
+    instances.each do |prov|
+      if resource = resources[prov.name]
+        resource.provider = prov
+      end
+    end
+  end
 
-def destroy
-  influx_cli(['-execute', "drop database #{resource[:name]}"].compact)
-end
+  def flush
+    if @property_flush[:ensure] == :absent
+        influx_cli(['-execute', "drop database #{resource[:name]}"].compact)
+        return
+    end
+    influx_cli(['-execute', "create database #{resource[:name]}"].compact)
+  end
+
+  def exists?
+    @property_hash[:ensure] == :present || false
+  end
+
+  def create
+    @property_flush[:ensure] = :present
+  end
+
+  def destroy
+    @property_flush[:ensure] = :absent
+  end
 
 end
